refactor(backend): move server key validation into core details
