+++ crates/services/miscellaneous/Cargo.toml
@@ -15,7 +15,6 @@ database-utils = { path = "../../utils/database" }
 dependent-models = { path = "../../models/dependent" }
 dependent-utils = { path = "../../utils/dependent" }
 enums = { path = "../../enums" }
-env-utils = { path = "../../utils/env" }
 futures = { workspace = true }
 itertools = { workspace = true }
 markdown = { workspace = true }
@@ -30,12 +29,10 @@ sea-orm = { workspace = true }
 sea-query = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
-serde_with = { workspace = true }
 supporting-service = { path = "../supporting" }
 tokio = { workspace = true }
 tracing = { workspace = true }
 traits = { path = "../../traits" }
-unkey = { workspace = true }
 user-models = { path = "../../models/user" }
 uuid = { workspace = true }
