+++ crates/services/supporting/Cargo.toml
@@ -8,10 +8,26 @@ async-graphql = { workspace = true }
 apalis = { workspace = true }
 background = { path = "../../background" }
 cache-service = { path = "../../services/cache" }
-common-models = { path = "../../models/common" }
+chrono = { workspace = true }
 chrono-tz = { workspace = true }
+common-models = { path = "../../models/common" }
+common-utils = { path = "../../utils/common" }
 config = { path = "../../config" }
+database-models = { path = "../../models/database" }
 dependent-models = { path = "../../models/dependent" }
+enums = { path = "../../enums" }
+env-utils = { path = "../../utils/env" }
 file-storage-service = { path = "../../services/file-storage" }
+isolang = { workspace = true }
+itertools = { workspace = true }
 openidconnect = { workspace = true }
+rustypipe = { workspace = true }
 sea-orm = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+serde_with = { workspace = true }
+tracing = { workspace = true }
+unkey = { workspace = true }
+
+[package.metadata.cargo-machete]
+ignored = ["tracing"]
